+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - improved documentation.
 - hedge now accepts ParseResults and can recursively add Hyperedge.text strings.
 - renamed function hyperedge.str2atom to str_to_atom.
+- Hyperedge is not a frozen dataclasse, instead of inheriting from tuple.
 
 ### Removed
 - function patterns.edge_matches_pattern.